+++ geometry.h
@@ -12,7 +12,7 @@
 
 void printRect(const glm::vec4& rect);
 int loadShaders(const GLchar* source, GLenum shaderType );
-double pointDistance(const glm::vec2& v1, const glm::vec2& v2);
+float pointDistance(const glm::vec2& v1, const glm::vec2& v2);
 glm::vec2 findMidpoint(const glm::vec4& v1); //midpoint of a line
 bool vecIntersect(const glm::vec4& vec1,const glm::vec4& vec2);
 bool vecIntersect(const glm::vec4& vec1,const glm::vec4& vec2, float angle1, float angle2); //vec1 and vec2 are rotated at the respective angles
